+++ packages/cli/src/commands/pull/index.ts
@@ -37,11 +37,9 @@ async function pullAllEnvFiles(
   return envPull(
     client,
     link,
-    project,
     environment,
     flags,
     [join('.vercel', environmentFile)],
-    client.output,
     cwd,
     'vercel-cli:pull'
   );