+++ packages/cli/src/util/env/get-env-records.ts
@@ -1,10 +1,7 @@
+import { URLSearchParams } from 'url';
 import { Output } from '../output';
 import Client from '../client';
-import type {
-  ProjectEnvVariable,
-  ProjectEnvTarget,
-} from '@vercel-internals/types';
-import { URLSearchParams } from 'url';
+import type { ProjectEnvVariable } from '@vercel-internals/types';
 
 /** The CLI command that was used that needs the environment variables. */
 export type EnvRecordsSource =
@@ -16,7 +13,6 @@ export type EnvRecordsSource =
   | 'vercel-cli:pull';
 
 export default async function getEnvRecords(
-  output: Output,
   client: Client,
   projectId: string,
   source: EnvRecordsSource,
@@ -25,11 +21,12 @@ export default async function getEnvRecords(
     gitBranch,
     decrypt,
   }: {
-    target?: ProjectEnvTarget | string;
+    target?: string;
     gitBranch?: string;
     decrypt?: boolean;
   } = {}
 ) {
+  const { output } = client;
   output.debug(
     `Fetching Environment Variables of project ${projectId} and target ${target}`
   );
@@ -62,7 +59,7 @@ export default async function getEnvRecords(
 }
 
 interface PullEnvOptions {
-  target?: ProjectEnvTarget | string;
+  target?: string;
   gitBranch?: string;
 }