+++ packages/frameworks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @vercel/frameworks
 
+## 3.2.0
+
+### Minor Changes
+
+- [fs-detectors] Adds new detector for Sanity v3 ([#11991](https://github.com/vercel/vercel/pull/11991))
+
 ## 3.1.1
 
 ### Patch Changes